+++ packages/client/.env
@@ -1,8 +1,4 @@
 NODE_ENV=development 
 CLIENT_PORT=3000
 VITE_API_URL='https://ya-praktikum.tech/api/v2'
-<<<<<<< 7.63
 VITE_HOST_URL='our_host'
-=======
-VITE_HOST_URL='our_host'
->>>>>>> main
