+++ packages/client/src/features/gameTimer/shared/formatTime.ts
@@ -1,5 +1,6 @@
 export const formatTime = (seconds: number) => {
   const mins = Math.floor(seconds / 60);
-  const formattedTime = `${mins.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+  const secs = seconds % 60;
+  const formattedTime = `${mins.toString().padStart(2, '0')}:${secs.toString().padStart(2, '0')}`;
   return formattedTime;
 };
